LLVM CorrectCode all working except For, Sementic analysis started, MIPS CorrectCode binaryOperation1.c and binaryOperation2.c

+++ tests/CompilersBenchmark/SemanticErrors/declarationDefinitionMismatch2.c
@@ -2,7 +2,7 @@
 int f(int a, int b);
 
 // definition of f with wrong parameter type
-int f(float a, int b){
+int f(int a, int b){
     return b;
 }
